+++ src/main/scala/io/viash/functionality/Functionality.scala
@@ -90,7 +90,7 @@ case class Functionality(
       "bash",
       "This results in the following output when calling the component with the `--help` argument:")
   @since("Viash 0.5.11")
-  @deprecated("Use `arguments` instead.", "Viash 0.6.0")
+  @deprecated("Use `arguments` instead.", "0.6.0", "0.8.0")
   inputs: List[Argument[_]] = Nil,
 
   @description("A list of output arguments in addition to the `arguments` list. Any arguments specified here will have their `type` set to `file` and thr `direction` set to `output` by default.")
@@ -111,7 +111,7 @@ case class Functionality(
       "bash",
       "This results in the following output when calling the component with the `--help` argument:")
   @since("Viash 0.5.11")
-  @deprecated("Use `arguments` instead.", "Viash 0.6.0")
+  @deprecated("Use `arguments` instead.", "0.6.0", "0.8.0")
   outputs: List[Argument[_]] = Nil,
   
   @description(
@@ -277,16 +277,16 @@ case class Functionality(
   // START OF REMOVED PARAMETERS THAT ARE STILL DOCUMENTED
   @description("Adds the resources directory to the PATH variable when set to true. This is set to false by default.")
   @since("Viash 0.5.5")
-  @removed("Extending the PATH turned out to be not desirable.", "Viash 0.5.11")
+  @removed("Extending the PATH turned out to be not desirable.", "", "0.5.11")
   private val add_resources_to_path: Boolean = false
 
   @description("One or more Bash/R/Python scripts to be used to test the component behaviour when `viash test` is invoked. Additional files of type `file` will be made available only during testing. Each test script should expect no command-line inputs, be platform-independent, and return an exit code >0 when unexpected behaviour occurs during testing.")
-  @deprecated("Use `test_resources` instead. No functional difference.", "Viash 0.5.13")
+  @deprecated("Use `test_resources` instead. No functional difference.", "0.5.13", "0.7.0")
   private val tests: List[Resource] = Nil
 
   @description("Setting this to false with disable this component when using namespaces.")
   @since("Viash 0.5.13")
-  @deprecated("Use `status` instead.", "Viash 0.6.0")
+  @deprecated("Use `status` instead.", "0.6.0", "0.8.0")
   private val enabled: Boolean = true
   // END OF REMOVED PARAMETERS THAT ARE STILL DOCUMENTED